Switch matmul pipeline to int8 inputs
Replace the demo and test matmul flow with integer semantics matching
the systolic array RTL: i8 inputs and i32 accumulation.

Update the MLIR demo, padding transform, systolic conversion pass, C
runtime ABI, socket protocol, Python simulator model, demo driver, and
matmul test templates to use i8/i8 -> i32 instead of f32.

+++ tests/matmul/cases.py
@@ -13,16 +13,16 @@ class MatmulTest:
   b_fn: Pattern
 
 
-def ones(i: int, j: int) -> float:
-  return 1.0
+def ones(i: int, j: int) -> int:
+  return 1
 
 
-def row_col(i: int, j: int) -> float:
-  return float((i + 2 * j) % 7 - 3)
+def row_col(i: int, j: int) -> int:
+  return (i + 2 * j) % 7 - 3
 
 
-def small_mod(i: int, j: int) -> float:
-  return float((3 * i - j) % 5 - 2)
+def small_mod(i: int, j: int) -> int:
+  return (3 * i - j) % 5 - 2
 
 
 TESTS = [
